security: add hsts header and tests
